move header to component
